--- gmd-crypto.js
@@ -1,369 +1,0 @@
-const axios = require('./get-axios');
-const cryptoUtil = require('./crypto-util');
-const KeyEncryption = require('./key-encryption');
-const Wallet = require('./lib/wallet');
-
-const GMD = {
-    baseURL: 'https://node.thecoopnetwork.io',
-    util: {},
-    Wallet: Wallet
-};
-
-/**
- *
- * @param {String} url of the GMD node. By default main net 'https://node.thecoopnetwork.io' is used.
- */
-GMD.setURL = (url) => {
-    GMD.baseURL = url;
-}
-
-/**
- *
- * @param {String} unsignedTransaction bytes as hex string.
- * @param {String} passPhrase usually 12 word passphrase
- * @returns [async] Signed transaction bytes. Signing is done locally (no passphrase is sent over noetwork)
- */
-GMD.signTransaction = async (unsignedTransaction, passPhrase) => {
-    let privateKey = await cryptoUtil.getPrivateKey(passPhrase);
-    return GMD.signTransactionPrivateKey(unsignedTransaction, privateKey);
-}
-
-/**
- * 
- * @param {string} message arbitrarary message; 
- * @param {*} privateKey private key in hex string format
- * @returns signature in hex string format
- */
-GMD.signHexMessagePrivateKey = async (message, privateKey) => {
-    return await cryptoUtil.signBytesPrivateKey(message, privateKey);
-}
-
-/**
- * Same as signHexMessagePrivateKey() except passphrase is received as param.
- * 
- * @param {string} message arbitrarary message in hex string format; 
- * @param {*} passPhrase usually 12 words
- * @returns signature in hex string format
- */
-GMD.signMessage = async (message, passPhrase) => {
-    let privateKey = await cryptoUtil.getPrivateKey(passPhrase);
-    return GMD.signHexMessagePrivateKey(message, privateKey);
-}
-
-/**
- * This call verifies a signature against a public key. Returns true if the message was signed by the 
- * corresponding private key.
- * 
- * @param {*} signature Signature in hex string format (same format as returned by GMD.signMessage())
- * @param {*} unsignedHexMessage Message in hex string format.
- * @param {*} publicKey Public key to check the signature.
- * @returns promise that resolves to boolean.
- */
-GMD.verifySignature = async (signature, unsignedHexMessage, publicKey) => {
-    return cryptoUtil.verifySignature(signature, unsignedHexMessage, publicKey);
-}
-
-/**
- *
- * @param {String} unsignedTransaction transaction bytes as hex string.
- * @param {String} privateKey hex string private key
- * @returns [async] Signed transaction bytes. Signing is done locally (no passphrase is sent over noetwork)
- */
-GMD.signTransactionPrivateKey = async (unsignedTransaction, privateKey) => {
-    const signature = await cryptoUtil.signBytesPrivateKey(unsignedTransaction, privateKey);
-    return unsignedTransaction.substr(0, 192) + signature + unsignedTransaction.substr(320);
-}
-
-/**
- * 
- * @param {string} str string
- * @returns hex string
- */
-GMD.util.strToHex = (str) => {
-    return cryptoUtil.strToHex(str);
-}
-
-GMD.util.encryptHex = async (messageHex, password) => {
-    return KeyEncryption.encryptHex(messageHex, password);
-}
-
-GMD.util.decryptToHex = async (encryptedJSON, password) => {
-    return KeyEncryption.decryptToHex(encryptedJSON, password);
-}
-
-/**
- *
- * @param {JSON} data Transaction data JSON input.
- * @returns boolean: true if JSON input contains properties "transactionJSON" and "unsignedTransactionBytes", false otherise.
- */
-GMD.isTransaction = (data) => {
-    return data && 'transactionJSON' in data && 'unsignedTransactionBytes' in data;
-}
-
-// const hasProperty = (obj, key) => {
-//     return Object.prototype.hasOwnProperty.call(obj, key);
-// }
-
-/**
- *
- * @param {JSON} data Transaction data JSON input.
- * @returns boolean: true if json represents transaction and contains "signatureHash" and "fullHash" properties.
- */
-GMD.isSignedTransactionResponse = (data) => {
-    return GMD.isTransaction(data) && 'signatureHash' in data && 'fullHash' in data;
-};
-
-/**
- *  API call to a GMD node. This call is done over network.
-
- *
- *
- * @param {String} method HTTP method (only 'get' or 'post' are used)
- * @param {JSON} params *  Full set of API methods and parameters can be seen here: https://node.thecoopnetwork.io/test.
- *  All parameters will be passed as a single json via 'params' parameter. Exception to this is 'secretPhrase' parameter which
- * should not be included in params, and even if you include it, the SDK will delete it before making the API call to the node.
- *  In addition to all parameters described above there are the following parameters:
- *  'requestType': [mandatory] wich is the name of the GMD API method (e.g. 'sendMoney', 'sendMessage', 'getPolls' etc..)
- *  'baseURL': [optional] URL of the GMD node where this request is performed. By default https://node.thecoopnetwork.io main net is used.
- *  'httpTimeout' [optional] parameter for HTTP request to specify a timeout when GMD node not reachable. Axios default is used if this param is not specified.
- *  Example:
- *  params = {
-        requestType: 'getAccountsBulk',
-        pageSize: 3,
-        page: 0,
-        baseURL: 'https://node.thecoopnetwork.io:6877'
-    }
- * @returns {Promise} that will resolve to the body of the server response (usually a JSON).
- */
-GMD.apiCall = async (method, params) => {
-    const { url, httpTimeout } = processParams(params);
-    const config = { method, url: url + '/nxt?' + (new URLSearchParams(params)).toString() };
-    if (httpTimeout && httpTimeout > 0) {
-        config.httpTimeout = httpTimeout;
-    }
-    return axios(config).then((res) => {
-        console.log(`Response status on request to ${config.url} is ${res.status}\nresponse body:\n${JSON.stringify(res.data, null, 2)}`);
-        return res.data;
-    })
-}
-
-/**
- * GMD.apiCallAndSign() does same thing as GMD.apiCall(). In addition, if the API call returns an unsigned transaction, this method
- * will sign it using the passPhrase param and broadcast it to the network. If the response does not contain an unsigned transaction,
- * or if password is invalid, nothing happens.
- * @param {String} method - samne as GMD.apiCall()
- * @param {JSON} params - samne as GMD.apiCall()
- * @param {String} passPhrase - secret passphrase, usually 12 words that will be used to sign the transaction (never sent over network)
- * @returns Promise that will resove to a JSON with the returned details of the broadcasted transaction.
- */
-GMD.apiCallAndSign = async (method, params, passPhrase) => {
-    let privateKey = await cryptoUtil.getPrivateKey(passPhrase);
-    return GMD.apiCallAndSignPrivateKey(method, params, privateKey);
-}
-
-/**
- * GMD.apiCallAndSignPrivateKey() does same thing as GMD.apiCall(). In addition, if the API call returns an unsigned transaction, this method
- * will sign it using the passPhrase param and broadcast it to the network. If the response does not contain an unsigned transaction,
- * or if password is invalid, nothing happens.
- * @param {String} method - samne as GMD.apiCall()
- * @param {JSON} params - samne as GMD.apiCall()
- * @param {String} privateKey - private key string in hex format
- */
-GMD.apiCallAndSignPrivateKey = async (method, params, privateKey) => {
-    const transaction = await GMD.apiCall(method, params);
-    if (GMD.isTransaction(transaction) && !GMD.isSignedTransactionResponse(transaction) && privateKey) {
-        const signedTransaction = await GMD.signTransactionPrivateKey(transaction.unsignedTransactionBytes, privateKey);
-        return GMD.broadcastSignedTransaction(signedTransaction);
-    }
-}
-
-const processParams = (params) => {
-    let url;
-    let httpTimeout;
-    if (params) {
-        if ('secretPhrase' in params) {
-            delete params.secretPhrase; // password is not sent to server - remove it from params - it is needed only to do local signing
-        }
-        if ('httpTimeout' in params) {
-            httpTimeout = params.httpTimeout;
-            delete params.httpTimeout;
-        }
-
-        if ('baseURL' in params) {
-            url = params.baseURL;
-            delete params.baseURL;
-        } else {
-            url = GMD.baseURL;
-        }
-    }
-    return { url, httpTimeout };
-}
-
-/**
- * Cryptographic transform of a secret to a public key. No request is sent, transformation happens locally.
- *
- * @param {String} pass  - secret passphrase, usually 12 words.
- * @returns - public key, hex string format
- */
-GMD.getPublicKey = async (pass) => {
-    return cryptoUtil.getPublicKey(pass);
-}
-
-// Helper functions
-
-/**
- * Helper function to broadcast a signed transaction
- * @param {String} signedTransaction The signed transaction bytes in hex string format.
- * @returns {JSON} The Node response on broadcast.
- */
-GMD.broadcastSignedTransaction = (signedTransaction) => {
-    return GMD.apiCall('post', { requestType: 'broadcastTransaction', transactionBytes: signedTransaction }).then(data => {
-        console.log('Succesfully posted the transaction broadcast. Data: ' + JSON.stringify(data, null, 2));
-    }).catch(err => {
-        console.log('Error posting transaction broadcast: ' + err);
-    });
-}
-
-/**
- *
- * @param {String} publicKey Public key in hex string format
- * @returns Promise that will resove to a json with account details: acount id, RS account id and public key.
- */
-GMD.getAccountId = (publicKey) => {
-    return GMD.apiCall('get', { requestType: 'getAccountId', publicKey });
-}
-
-/**
- *  Getting public key from RS account is not cryptographically possible.
- * However, the mapping RS account <-> public key is available on the node if this account exists on the chain
- * (if at least one transaction has been made).
- */
-GMD.getPublicKeyFromRS = (rsAccount) => {
-    return GMD.apiCall('get', { requestType: 'getAccountPublicKey', account: rsAccount }).then(data => {
-        return data.publicKey;
-    });
-}
-
-GMD.getWalletFromEncryptedJSON = async (encryptedJSON, encryptionPassword) => {
-    return Wallet.walletFromEncryptedJSON(encryptedJSON, encryptionPassword);
-};
-
-GMD.generateWallet = async () => {
-    let secretPassphrase = require('./pass-gen').generatePass();
-    return GMD.generateWalletFromPassphrase(secretPassphrase);
-}
-
-GMD.generateWalletFromPassphrase = async (secretPassphrase) => {
-    return Wallet.fromPassphrase(secretPassphrase);
-};
-
-GMD.getEncryptedJSONFromPassphrase = async (passPhrase, encryptionPassword) => {
-    Wallet.encryptedJSONFromPassPhrase(passPhrase, encryptionPassword);
-}
-
-/**
- * Generate full GMD account.
- * @param {String} secretPassphrase [optional ]if not provided, a secret 12 word passphrase will be generated. Generation of passphrase happens locally,
- * it is never sent on the network. Password is transformed to a public key and that key is sent to a node to get the account id details.
- * @returns a promise that resolves to a JSOM containing: account ID, RS account ID (format GMD-...), public key, secret passphrase.
- */
-GMD.generateAccount = async (secretPassphrase) => {
-    if (!secretPassphrase) {
-        const PassPhraseGenerator = require('./pass-gen');
-        secretPassphrase = PassPhraseGenerator.generatePass();
-    }
-    return GMD.getWalletDetailsFromPassPhrase(secretPassphrase);
-}
-
-/**
- *
- * @param {*} secretPassphrase is transformed to a public key and that key is sent to a node to get the account id details.
- * @returns a promise that resolves to a JSON containing: account ID, RS account ID (format GMD-...), public key, private key, secret passphrase.
- */
-GMD.getWalletDetailsFromPassPhrase = async (secretPassphrase) => {
-    let { publicKey, privateKey } = await cryptoUtil.getPublicPrivateKey(secretPassphrase);
-
-    return GMD.getAccountId(publicKey).then((data) => {
-        data.secretPassphrase = secretPassphrase;
-        data.privateKey = privateKey;
-        return data;
-    })
-}
-
-/**
- *
- * @param {*} rsAccount
- * @returns true if RS account checsum and format is checks out, false is invalid address. Actual check is done remotely on a Coop Network Node, so lack of connectivity will result in error thrown.
- */
-GMD.checkRSAddress = async (rsAccount) => {
-    return GMD.apiCall('get', { requestType: 'rsConvert', account: rsAccount }).then(data => {
-        return 'accountLongId' in data;
-    })
-}
-
-/**
- * Returns a promise that will resolve to a balanceNQT
- */
-GMD.getBalance = async (rsAccount) => {
-    return GMD.apiCall('get', { requestType: 'getBalance', account: rsAccount }).then(data => {
-        return data.balanceNQT;
-    })
-}
-
-/**
- * Sends GMD to another wallet. The implementation does 3 operations:
- * 1. Creates an unsigned transaction for sending GMD by calling appropriate API on the node.
- * 2. Waits for the transaction created on step 1 to be received from the node and signs it locally.
- * 3. Broadcasts to the GMD nodes the signed transaction.
- *
- * @param {String} recipient RS destination account ('GMD-...' format)
- * @param {String} amountNQT Amount to be transfered in NQT. 1 GMD = 100000000 NQT.
- * @param {String} passPhrase Secret passphrase, used for step 2, local signing.
- * @param {String} feeNQT Fee in NQT
- * @returns Promise that resolves to a JSON with broadcast restult.
- */
-GMD.sendMoney = (recipient, amountNQT, passPhrase, feeNQT) => {
-    return sendMoneyAPICall(true, recipient, amountNQT, passPhrase, feeNQT);
-}
-
-/**
- * Same as GMD.sendMoney(), but onyl step 1. This is usefull when the signing and broadcast is done on other device (e.g. mobile).
- * Sender is identified thorough its public key, not secret passphrase.
- * @param {String} recipient same as GMD.sendMoney()
- * @param {String} amountNQT same as GMD.sendMoney()
- * @param {String} senderPublicKey sender public key.
- * @param {String} feeNQT same as GMD.sendMoney()
- * @returns Promise that resolves to a transaction JSON. If signing on other device, create a QR code with the string found in
- * transaction.unsignedTransactionBytes property. This string may be signed on the other device with GMD.signTransaction() (assuming
- * the other device uses this SDK)
- */
-GMD.createUnsignedSendMoneyTransaction = (recipient, amountNQT, senderPublicKey, feeNQT) => {
-    return sendMoneyAPICall(false, recipient, amountNQT, senderPublicKey, feeNQT);
-}
-
-const sendMoneyAPICall = async (signIt, recipient, amountNQT, from, feeNQT) => {
-    const params = {
-        requestType: 'sendMoney',
-        recipient,
-        amountNQT,
-        feeNQT,
-        deadline: '1440'
-    };
-
-    if (signIt) {
-        params.publicKey = await GMD.getPublicKey(from);
-    } else {
-        params.publicKey = from;
-    }
-    const transaction = await GMD.apiCall('post', params);
-    if (signIt) {
-        if (GMD.isTransaction(transaction) && !GMD.isSignedTransactionResponse(transaction) && from) {
-            const signedTransaction = await GMD.signTransaction(transaction.unsignedTransactionBytes, from);
-            return GMD.broadcastSignedTransaction(signedTransaction);
-        }
-    } else {
-        return transaction;
-    }
-}
-
-module.exports = GMD;
