--- index.js
@@ -1,8 +1,0 @@
-const GMD = require('./gmd-crypto');
-const GMDEvents = require('./gmd-events');
-
-
-module.exports = {
-    GMD: GMD,
-    GMDEvents: GMDEvents
-}
