+++ lib/crypto-util.js
@@ -1,11 +1,11 @@
-const crypto = require('./get-crypto');
+const webcrypto = require('./get-crypto');
 const curve25519 = require('./curve25519');
 const cryptoUtil = {};
 
 
 cryptoUtil.strToHex = (str) => {
     let result = '';
-    cryptoUtil.strToBytes(str).forEach(c=> result+=c.toString(16));
+    cryptoUtil.strToBytes(str).forEach(c => result += c.toString(16));
     return result;
 }
 
@@ -14,7 +14,7 @@ cryptoUtil.strToBytes = (str) => {
     for (let i = 0; i < str.length; i++) {
         result.push(str.charCodeAt(i));
     }
-    return result;   
+    return result;
 }
 
 cryptoUtil.hexToString = (hex) => {
@@ -100,7 +100,7 @@ cryptoUtil.SHA256 = async (in1, in2) => {
     }
 
     let arrayBufferInput = Uint8Array.from(input);
-    let output = await crypto.subtle.digest('SHA-256', arrayBufferInput);
+    let output = await webcrypto.subtle.digest('SHA-256', arrayBufferInput);
     return Array.from(new Uint8Array(output));
 }
 
@@ -138,7 +138,7 @@ cryptoUtil.wordsToBytes = (wordArr) => {
 
 
 
-cryptoUtil.signBytes = async (message, passPhrase)=>{
+cryptoUtil.signBytes = async (message, passPhrase) => {
     let privateKey = await cryptoUtil.getPrivateKey(passPhrase);
     return cryptoUtil.signBytesPrivateKey(message, privateKey);
 }
@@ -158,17 +158,17 @@ cryptoUtil.verifySignature = async (signature, unsignedMessage, publicKey) => {
     let signatureBytes = cryptoUtil.hexToBytes(signature);
     let messageBytes = cryptoUtil.hexToBytes(unsignedMessage);
     let publicKeyBytes = cryptoUtil.hexToBytes(publicKey)
-    let v = signatureBytes.slice(0,32);
+    let v = signatureBytes.slice(0, 32);
     let h = signatureBytes.slice(32);
     let Y = curve25519.verify(v, h, publicKeyBytes);
 
     let m = await cryptoUtil.SHA256(messageBytes);
     let h2 = await cryptoUtil.SHA256(m, Y);
 
-    return cryptoUtil.byteArraysEqual(h,h2);
+    return cryptoUtil.byteArraysEqual(h, h2);
 }
 
-cryptoUtil.byteArraysEqual  = (bytes1, bytes2) => {
+cryptoUtil.byteArraysEqual = (bytes1, bytes2) => {
     if (bytes1.length !== bytes2.length) {
         return false;
     }
@@ -182,12 +182,12 @@ cryptoUtil.byteArraysEqual  = (bytes1, bytes2) => {
 
 
 cryptoUtil.getPrivateKey = async (pass) => {
-    let {privateKey} = await cryptoUtil.getPublicPrivateKey(pass);
+    let { privateKey } = await cryptoUtil.getPublicPrivateKey(pass);
     return privateKey;
 }
 
 cryptoUtil.getPublicKey = async (pass) => {
-    let {publicKey} = await cryptoUtil.getPublicPrivateKey(pass);
+    let { publicKey } = await cryptoUtil.getPublicPrivateKey(pass);
     return publicKey;
 }
 
@@ -198,8 +198,8 @@ cryptoUtil.getPublicPrivateKey = async (passPhrase) => {
 
 
 cryptoUtil.getPublicPrivateKeyFromSeed = (seed) => {
-    let {p,s} = curve25519.keygen(seed);
-    return {publicKey: cryptoUtil.bytesToHex(p), privateKey: cryptoUtil.bytesToHex(s) }
+    let { p, s } = curve25519.keygen(seed);
+    return { publicKey: cryptoUtil.bytesToHex(p), privateKey: cryptoUtil.bytesToHex(s) }
 }
 
 cryptoUtil.getSeed = async (passPhrase) => {
