+++ lib/curve25519.js
