+++ lib/get-axios.js
@@ -4,3 +4,5 @@
  * Swapping is configured in packge.json
  */
 module.exports = require('axios');
+// /* eslint-disable no-undef */
+// module.exports = axios;
