+++ lib/get-crypto.js
@@ -4,3 +4,4 @@
  * Swapping is configured in packge.json
  */
 module.exports = require('crypto').webcrypto; //Node equivalent of window.crypto in browser
+//module.exports = window.crypto;
