+++ lib/key-encryption.js
