+++ lib/pass-gen.js
